refactor: change unite legal name and entreprise service name
