+++ packages/api/src/modules/providers/uniteLegalEntreprises/uniteLegal.entrepises.service.ts
@@ -16,7 +16,8 @@ export class UniteLegalEntreprisesService {
         }
     }
 
-    async isEntreprise(siren: Siren) { //sirenIsEntreprise
+    async isEntreprise(siren: Siren) {
+        //sirenIsEntreprise
         return !!(await uniteLegalEntreprisePort.findOneBySiren(siren));
     }
 }
